Outlined the logic for the Speech Recognition API.
Adjust button text on enable, record, and stop recording states.

Wrote state variable and handler functions for start and stop recording.

Decided to factor out the API interaction logic into a seperate file (Not included in this commit) and imported in ChatBot.js as an JSON object with a field for a method to update the transcription text for the ChatBot.js file